cuda: narrow the stored Q to f16 (L045)
PULSAR_Q_ELT_SIZE 4u -> 2u and pulsar_q_t float -> __half.  Every consumer was
templated over the preceding commits, so this flip is the two definitions plus
the three things a clean compile could not have told us.

The source model carries E4M3 activations; f32 Q was over-precision we were
paying for in bandwidth.  Per the note on PULSAR_Q_ELT_SIZE this is NOT
bit-exact -- under q_prep_active batch_q holds unnormalised q_b output and the
per-head sum of squares is taken from it, so narrowing moves the inputs to a
reduction, not just its storage.  cuda-reference-gate is what certifies it.

Three defects the compiler could not see, all of the same shape -- a
pulsar_gpu_tensor carries no element type, so reinterpreting its bytes is
type-legal:

  - Both Q producers still asked the mxfp8 GEMM for an f32 output: gpu_decode
    passed the literal 0, and gpu_graph_matmul_mxfp8_named_tensor hardcoded 0
    for all six of its callers.  Against the narrowed buffer that is out_dim*4
    bytes into an out_dim*2 allocation -- a 2x overrun AND every Q value wrong.
    The wrapper now takes out_f16 per call site, and the Q sites derive it from
    PULSAR_Q_ELT_SIZE rather than writing a literal, because those are the same
    fact and this is what it costs when they drift.

  - The Q debug dumps read n*sizeof(float) from the same buffer -- the 2x
    out-of-bounds read the HC carrier flip had already hit and solved.  Q now
    has the matching reader and dump, and stays out of the f32 range-sweep
    census, whose question it no longer poses.

  - Nothing stopped the GEMM writing past its destination, so add that bound.
    It is a bounds check rather than a format check: batch buffers are sized
    for full prefill capacity, so it catches every decode-shaped case and the
    large batches, but cannot see a mismatch that still fits.  Stated in the
    comment so the next reader does not over-trust it.

Verified: full build clean, seam-check 81 host files CUDA-API clean.  The
model-dependent gates need sparky and have not run yet.

+++ src/cuda/pulsar_cuda_internal.h
@@ -110,7 +110,7 @@ static inline int pulsar_env_tier_on(const char *name) {
 
 /* Stored Q element type; pairs with PULSAR_Q_ELT_SIZE in pulsar_gpu.h.
  * float today, __half at the flip. */
-typedef float pulsar_q_t;
+typedef __half pulsar_q_t;
 
 /* The cudaDataType the stored Q presents to cuBLAS.  Derived from pulsar_q_t
  * rather than written out, so a GEMM's operand type can never drift from the
